+++ appliance_settings_test.go
@@ -1,0 +1,14 @@
+package morpheus_test
+
+import (
+	"testing"
+
+	"github.com/gomorpheus/morpheus-go-sdk"
+)
+
+func TestGetApplianceSettings(t *testing.T) {
+	client := getTestClient(t)
+	req := &morpheus.Request{}
+	resp, err := client.GetApplianceSettings(req)
+	assertResponse(t, resp, err)
+}