+++ examples/license/main.go
@@ -16,12 +16,12 @@ func main() {
 	}
 	fmt.Println("LOGIN RESPONSE:", resp)
 
-	// List cluster layouts
+	// Get license
 	req := &morpheus.Request{}
-	response, err := client.ListClusterLayouts(req)
+	licenseResponse, err := client.GetLicense(req)
 	if err != nil {
 		log.Fatal(err)
 	}
-	result := response.Result.(*morpheus.ListClusterLayoutsResult)
-	log.Println(result.ClusterLayouts)
+	result := licenseResponse.Result.(*morpheus.GetLicenseResult)
+	log.Println(result)
 }