+++ provisioning_settings.go
@@ -1,0 +1,69 @@
+package morpheus
+
+var (
+	// ProvisioningSettingsPath is the API endpoint for provisioning settings
+	ProvisioningSettingsPath = "/api/provisioning-settings"
+)
+
+// ProvisioningSettings structures for use in request and response payloads
+type ProvisioningSettings struct {
+	AllowZoneSelection          bool   `json:"allowZoneSelection"`
+	AllowServerSelection        bool   `json:"allowServerSelection"`
+	RequireEnvironments         bool   `json:"requireEnvironments"`
+	ShowPricing                 bool   `json:"showPricing"`
+	HideDatastoreStats          bool   `json:"hideDatastoreStats"`
+	CrossTenantNamingPolicies   bool   `json:"crossTenantNamingPolicies"`
+	ReuseSequence               bool   `json:"reuseSequence"`
+	ShowConsoleKeyboardSettings bool   `json:"showConsoleKeyboardSettings"`
+	CloudInitUsername           string `json:"cloudInitUsername"`
+	CloudInitPassword           string `json:"cloudInitPassword"`
+	CloudInitPasswordHash       string `json:"cloudInitPasswordHash"`
+	Cloudinitkeypair            struct {
+		ID   int    `json:"id"`
+		Name string `json:"name"`
+	} `json:"cloudInitKeyPair"`
+	WindowsPassword     string      `json:"windowsPassword"`
+	WindowsPasswordHash string      `json:"windowsPasswordHash"`
+	PXERootPassword     interface{} `json:"pxeRootPassword"`
+	PXERootPasswordHash interface{} `json:"pxeRootPasswordHash"`
+	DefaultTemplateType struct {
+		ID   int    `json:"id"`
+		Name string `json:"name"`
+		Code string `json:"code"`
+	} `json:"defaultTemplateType"`
+	DeployStorageProvider struct {
+		ID   int    `json:"id"`
+		Name string `json:"name"`
+	} `json:"deployStorageProvider"`
+}
+
+type GetProvisinoingSettingsResult struct {
+	ProvisioningSettings *ProvisioningSettings `json:"provisioningSettings"`
+}
+
+type UpdateProvisioningSettingsResult struct {
+	Success              bool                  `json:"success"`
+	Message              string                `json:"msg"`
+	Errors               map[string]string     `json:"errors"`
+	ProvisioningSettings *ProvisioningSettings `json:"provisioningSettings"`
+}
+
+// Client request methods
+func (client *Client) GetProvisioningSettings(req *Request) (*Response, error) {
+	return client.Execute(&Request{
+		Method:      "GET",
+		Path:        ProvisioningSettingsPath,
+		QueryParams: req.QueryParams,
+		Result:      &GetProvisinoingSettingsResult{},
+	})
+}
+
+func (client *Client) UpdateProvisioningSettings(req *Request) (*Response, error) {
+	return client.Execute(&Request{
+		Method:      "PUT",
+		Path:        ProvisioningSettingsPath,
+		QueryParams: req.QueryParams,
+		Body:        req.Body,
+		Result:      &UpdateProvisioningSettingsResult{},
+	})
+}